Update for Windows 10.1511 ==========================
VBS checks for installed voices and substitutes a compatible voice
if the first choice for the region is not available.  This fixes a
behaviour where multilingual users using an English system might
hear an English voice if the language and region did not match
exactly.  For example, with an es-MX voice installed, es-ES text
would be pronounced using the English default Microsoft David
voice instead of the Mexican Spanish Sabrina voice.