Allow custom environment variable =================================
The extension may use a system temporary directory to
write temporary files like lock files.  On some systems,
the default temporary directory may be locked to users
or processes that do not have sufficient reading or writing
privileges.

The preferred way of looking up a temporary directory is
to use an environmental variable.  If you have a custom
platform, you can set the `READTEXTTEMP` variable using
the `env` or `set` command in a terminal to prioritize
a specific directory as the temporary directory.  If the
`READTEXTTEMP` variable is set, then it is the first
place that the extension looks to read and write temporary
files.